+++ tests/e2e/core-tests/specs/index.js
@@ -26,6 +26,8 @@ const runTaxSettingsTest = require( './merchant/wp-admin-settings-tax.test' );
 const runOrderStatusFiltersTest = require( './merchant/wp-admin-order-status-filters.test' );
 const runOrderRefundTest = require( './merchant/wp-admin-order-refund.test' );
 const runOrderApplyCouponTest = require( './merchant/wp-admin-order-apply-coupon.test' );
+const runProductEditDetailsTest = require( './merchant/wp-admin-product-edit-details.test' );
+const runProductSearchTest = require( './merchant/wp-admin-product-search.test' );
 const runMerchantOrdersCustomerPaymentPage = require( './merchant/wp-admin-order-customer-payment-page.test' );
 
 const runSetupOnboardingTests = () => {
@@ -55,6 +57,8 @@ const runMerchantTests = () => {
 	runOrderStatusFiltersTest();
 	runOrderRefundTest();
 	runOrderApplyCouponTest();
+	runProductEditDetailsTest();
+	runProductSearchTest();
 	runMerchantOrdersCustomerPaymentPage();
 }
 
@@ -81,6 +85,8 @@ module.exports = {
 	runOrderStatusFiltersTest,
 	runOrderRefundTest,
 	runOrderApplyCouponTest,
+	runProductEditDetailsTest,
+	runProductSearchTest,
 	runMerchantOrdersCustomerPaymentPage,
 	runMerchantTests,
 };